requirements-dev.txt
-r requirements.txt

pytest==8.3.3
pytest-dotenv==0.5.2
ruff==0.7.3
sqlalchemy_utils==0.41.2
httpx==0.27.2


